Add server mode
Server mode will sleep after finishing its download, enter update mode,
then proceed to loop back around and re-scan the selected blogs. It can
be toggled with the `-server` flag.

This is mostly for convenience and will probably be unsupported, since
this tool was made to be run occasionally instead of constantly.

Resolves #13.