Add hardlinking to duplicate files (resolves #29)
Hardlinking files will make two of the same files point to the same
file instead of downloading two different files and taking two times
the needed space (more if multiple blogs reblogged it). The files are
hardlinked based on name, not on content. This may change in the
future, but it may require a lot more memory and processing power than
I want to use for now, at least with the naive implementation.

This is backwards-compatible: on startup, it analyzes and hardlinks
files that should be the same, based on name. This saves space right
away.

This solution unfortunately begins taking up a lot of memory after a
certain point, due to keeping track of every single file in the
download directory. It needs to be refactored and possibly improved.

There are a few bugs that occur when files are deleted during download
and they are the primary file being pointed to by the file tracker.
This possibly needs to be addressed, but is not too urgent.

This is also obviously harmful if run in directories that aren't
dedicated download directories, due to the hardlinker checking for
filenames instead of actual contents. For instance, it could link
two system files with the same name, but different functionality,
together.

eggs, milk, bread, butter, instant coffee