Don't check for file similarity on startup scan (fixes #39)
This avoids even more needless stat calls.

This can probably be sped up even more if the directory scan is run in
parallel, and separately from adding files to the existing file list.
Currently, it's set up so that it scans a directory, then adds the
files to the list, then repeats. Ideally, it could scan a directory,
hand off the list to a processor, then continue scanning, while the
processor adds the filenames to the list.

However, this is probably not necessary, since getting the info from
the directory is still the bottleneck. Implementing this would speed
the startup process up by a few seconds per hundred blogs, at most.

Another possible thing to look into to speed the process up: actively scrape
users while the directory scan is happening, and start downloading files
when it's done. Yay, parallelism.