Correct the usage of concurrent maps
In Go, maps are not safe for concurrent access. Yet the code accesses
the filetracker map concurrently in several places. Newer versions of
Go detects this and panic the program if this happens.

Use the existing mutex lock to protect access to the map. It is highly
unlikely to create any kind of contention as most of the programs
contention is centered around waiting on files to download over the
network.