Fix concurrent accesses to counters
Running the program under the race detector (go install -race)
uncovers a number of data races in the program. These all have to do
with the situation where one CPU core is writing, through an atomic
operation, to a variable, yet other cores access that variable
nonatomically. This isn't safe in the Go memory guarantees.

Alter the code to use the atomic.LoadUint64 operation for these
accesses which protect the program from data races and makes the
program able to run under the race detector with no errors mentioned.